requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
asyncpg>=0.30.0
openai>=1.0.0

